requirements.txt
python-docx>=1.1.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
httpx>=0.27.0
pytest>=7.0.0
pytest-mock>=3.10.0
